+++ src/components/Breadcrumbs.vue
@@ -3,8 +3,8 @@
     <template #item="{ item }">
       <v-breadcrumbs-item
         class="pa-0"
-        :title="$t(`Router.${item.title}.Title`)"
-        :to="item.to"
+        :title="getItemTitle(item)"
+        :to="getItemTo(item)"
         :disabled="item.disabled"
       />
     </template>
@@ -16,7 +16,25 @@ export default {
   computed: {
     breadcrumbs() {
       return this.$route.meta.breadcrumbs
+    },
+    objectId() {
+      return this.$route.params.id
+    },
+    objectUsername() {
+      return this.$route.params.username
     }
   },
+  methods: {
+    getItemTitle(item) {
+      if (this.objectId && item.to && item.to.includes(':id')) return this.objectId
+      if (this.objectUsername && item.to && item.to.includes(':username')) return this.objectUsername
+      return this.$t(`Router.${item.title}.Title`)
+    },
+    getItemTo(item) {
+      if (this.objectId && item.to && item.to.includes(':id')) return item.to.replace(':id', this.objectId)
+      if (this.objectUsername && item.to && item.to.includes(':username')) return item.to.replace(':username', this.objectUsername)
+      return item.to
+    }
+  }
 }
 </script>