+++ src/components/UserActionMenuButton.vue
@@ -8,6 +8,9 @@
         </v-list-subheader>
         <v-divider />
         <ShareLink :overrideUrl="getShareLinkUrl" display="list-item" />
+        <v-list-item :slim="true" prepend-icon="mdi-image" :to="getUserProofListUrl">
+          {{ $t('Common.Proofs') }}
+        </v-list-item>
         <OpenFoodFactsLink facet="editor" :value="user.user_id" display="list-item" />
       </v-list>
     </v-menu>
@@ -50,6 +53,9 @@ export default {
     },
     getShareLinkUrl() {
       return `/users/${this.user.user_id}`
+    },
+    getUserProofListUrl() {
+      return `/users/${this.user.user_id}/proofs`
     }
   }
 }